Changes to stop sql errors
