+++ src/test/resources/json/addfilemetadata_data_fileid_not_exists.json
@@ -1,2 +1,2 @@
-{"data": null,"errors": [{ "message": "Referential integrity constraint violation: \"CONSTRAINT_A36: PUBLIC.FileMetadata FOREIGN KEY(FileId) REFERENCES PUBLIC.File(FileId) ('28f78f42-0f77-4384-bcdb-04d85100171e')\"; SQL statement:\ninsert into \"FileMetadata\" (\"MetadataId\",\"FileId\",\"PropertyId\",\"Value\",\"Datetime\",\"UserId\")  values (?,?,?,?,?,?) [23506-200]", "locations": [{"column": 47,"line": 1} ]}]}
+{"data": null,"errors": [{ "message": "Could not find metadata for file 28f78f42-0f77-4384-bcdb-04d85100171e", "locations": [{"column": 47,"line": 1} ]}]}
