+++ src/test/scala/uk/gov/nationalarchives/tdr/api/routes/ClientFileMetadataRouteSpec.scala
@@ -153,24 +153,6 @@ class ClientFileMetadataRouteSpec extends AnyFlatSpec with Matchers with TestReq
     ps.executeUpdate()
   }
 
-  // scalastyle:off magic.number
-  private def createClientFileMetadata(fileId: UUID): Unit = {
-    val sql = s"insert into ClientFileMetadata (FileId,OriginalPath,Checksum,ChecksumType,LastModified,CreatedDate,Filesize,Datetime,ClientFileMetadataId) " +
-      s"VALUES (?,?,?,?,?,?,?,?,?)"
-    val ps: PreparedStatement = DbConnection.db.source.createConnection().prepareStatement(sql)
-    ps.setString(1, fileId.toString)
-    ps.setString(2, "originalPath")
-    ps.setString(3, "checksum")
-    ps.setString(4, "checksumType")
-    ps.setTimestamp(5, Timestamp.from(FixedTimeSource.now))
-    ps.setTimestamp(6, Timestamp.from(FixedTimeSource.now))
-    ps.setString(7, "1")
-    ps.setTimestamp(8, Timestamp.from(FixedTimeSource.now))
-    ps.setString(9, UUID.randomUUID.toString)
-    ps.executeUpdate()
-  }
-  // scalastyle:on magic.number
-
   private def checkClientFileMetadataExists(clientFileMetadataId: UUID): Unit = {
     val sql = "select * from ClientFileMetadata where ClientFileMetadataId = ?;"
     val ps: PreparedStatement = DbConnection.db.source.createConnection().prepareStatement(sql)
