+++ src/test/scala/uk/gov/nationalarchives/tdr/api/utils/TestUtils.scala
@@ -90,6 +90,7 @@ object TestUtils {
     val consignmentId = UUID.fromString("eb197bfb-43f7-40ca-9104-8f6cbda88506")
     createConsignment(consignmentId, userId)
     createFile(defaultFileId, consignmentId)
+    createClientFileMetadata(defaultFileId)
   }
 
   def createConsignment(consignmentId: UUID, userId: UUID): Unit = {
@@ -152,6 +153,22 @@ object TestUtils {
     ps.executeUpdate()
   }
 
+  def createClientFileMetadata(fileId: UUID): Unit = {
+    val sql = s"insert into ClientFileMetadata (FileId,OriginalPath,Checksum,ChecksumType,LastModified,CreatedDate,Filesize,Datetime,ClientFileMetadataId) " +
+      s"VALUES (?,?,?,?,?,?,?,?,?)"
+    val ps: PreparedStatement = DbConnection.db.source.createConnection().prepareStatement(sql)
+    ps.setString(1, fileId.toString)
+    ps.setString(2, "originalPath")
+    ps.setString(3, "checksum")
+    ps.setString(4, "checksumType")
+    ps.setTimestamp(5, Timestamp.from(FixedTimeSource.now))
+    ps.setTimestamp(6, Timestamp.from(FixedTimeSource.now))
+    ps.setString(7, "1")
+    ps.setTimestamp(8, Timestamp.from(FixedTimeSource.now))
+    ps.setString(9, UUID.randomUUID.toString)
+    ps.executeUpdate()
+  }
+
   //scalastyle:on magic.number
 
   def addFileProperty(propertyId: String, name: String): Unit = {
@@ -162,5 +179,6 @@ object TestUtils {
 
     ps.executeUpdate()
   }
+
 
 }
